+++ plugins/xlsx-extractor/CHANGELOG.md
@@ -1,4 +1,12 @@
-# @flatfile/plugin-record-hook
+# @flatfile/plugin-xlsx-extractor
+
+## 1.9.1
+
+### Patch Changes
+
+- c39cc51: Fix for xlsx extraction when file contains empty sheets
+- Updated dependencies [c39cc51]
+  - @flatfile/util-extractor@0.4.9
 
 ## 1.9.0